renamed gatekeeper package to avoid confusion with the gatekeepr app directory.
